+++ bambi/tests/test_model_construction.py
@@ -257,3 +257,31 @@ def test_categorical_term():
 
     for term, expected in zip(terms, expecteds):
         assert model.terms[term].categorical is expected
+
+
+def test_omit_offsets_false():
+    data = pd.DataFrame(
+        {
+            "y": np.random.normal(size=100),
+            "x1": np.random.normal(size=100),
+            "g1": ["a"] * 50 + ["b"] * 50,
+        }
+    )
+    model = Model(data)
+    fitted = model.fit("y ~ x1", group_specific=["x1|g1"], omit_offsets=False)
+    offsets = [v for v in fitted.posterior.dims if "offset" in v]
+    assert offsets == ["1|g1_offset_dim_0", "x1|g1_offset_dim_0"]
+
+
+def test_omit_offsets_true():
+    data = pd.DataFrame(
+        {
+            "y": np.random.normal(size=100),
+            "x1": np.random.normal(size=100),
+            "g1": ["a"] * 50 + ["b"] * 50,
+        }
+    )
+    model = Model(data)
+    fitted = model.fit("y ~ x1", group_specific=["x1|g1"], omit_offsets=True)
+    offsets = [v for v in fitted.posterior.dims if "offset" in v]
+    assert not offsets